refactor: simplify transfer-eth-erc20/to-near
`proof_locker` and `proof` were calculated and saved during discrete
steps of the transfer process, rather than as-needed when minting. This
made it hard to understand where the data was coming from and why it was
being calculated while reading through the code.

Additionally, there was a fair bit of indirection and awkward code
throughout. This has been cleaned up.